requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
email-validator==2.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
mysql-connector-python==8.1.0
python-dotenv==1.0.0
